Move hub into .hub.hub
