+++ homeassistant/components/unifi/hub/__init__.py
@@ -1,0 +1,3 @@
+"""Internal functionality not part of HA infrastructure."""
+
+from .hub import UnifiHub, get_unifi_api  # noqa: F401
